+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Unreleased
 
-- [Data Frame](https://inspect.aisi.org.uk/dataframe.html) functions for reading data frames from log files.
+- Added support for [Tavily](https://tavily.com/)'s Research API to the `web_search` tool.
+- [DataFrame](https://inspect.aisi.org.uk/dataframe.html) functions for reading dataframes from log files.
 - Multiple choice: `max_tokens` option to control tokens used for `generate()`.
 - Don't enforce sample `working_limit` after solvers have completed executing (matching behavior of other sample limits).
 - Only pass `user` parameter on to sandboxes if is not `None` (eases compatibility with older sandbox providers).
 - Anthropic: Retry when `type` in the error message body is "overloaded_error". 
+- Model Context Protocol: Update to typing changes made in v1.8.0 of `mcp` package (now the minimum required version).
 - TaskState: `input_text` and `user_prompt` properties now read the last rather than first user message.
 - Inspect View: Properly display 'more' options when content is collapsed.
 - Inspect View: Fix issue that prevented filtering of sample list when viewing a running evaluation.